deviceproject/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
Django==3.0.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djongo==1.3.3
idna==2.10
Naked==0.1.31
Pillow==8.0.1
pyasn1==0.4.8
pycrypto==2.6.1
pycryptodome==3.9.9
pycryptodomex==3.9.9
PyJWT==1.7.1
pymongo==3.11.2
pytz==2020.4
PyYAML==5.3.1
qrcode==6.1
requests==2.25.0
rsa==4.6
shellescape==3.8.1
six==1.15.0
sqlparse==0.2.4
urllib3==1.26.2
